rename: 컴포넌트 이름 수정 및 폴더 이동
