+++ src/app/_components/molecules/ModalContent/postFeed/ImagePreview/ImagePreview.tsx
@@ -1,22 +1,22 @@
 import { ModalStyle } from "@components/atoms/modal/Modal.style";
 import React from "react";
-import { Content } from "./EditImageModal.style";
+import { Content } from "./ImagePreview.style";
 import BaseButton from "@components/atoms/button/BaseButton";
 import { faArrowLeft } from "@fortawesome/free-solid-svg-icons";
 import useModalStore from "@/_stores/client/modalStore";
 import Image from "next/image";
 import IconButton from "@components/atoms/button/IconButton";
-import { MODAL_NAME } from "@/_constant/modal";
+import { MODAL } from "@/_constant/modal";
 
-export default function EditImageModal() {
-  const { modalName, selectedImage, setModal } = useModalStore();
+export default function ImagePreview() {
+  const { selectedImage, setModalStep } = useModalStore();
   const handleNextClick = () => {
-    setModal(MODAL_NAME.WRITE_POST);
+    setModalStep(MODAL.POST_FEED.STEP.WRITE_CONTENT);
   };
   const handleBackClick = () => {
     const res = confirm("게시물을 삭제하시겠어요?").valueOf();
     if (res) {
-      setModal(MODAL_NAME.POST_FEED);
+      setModalStep(MODAL.POST_FEED.STEP.UPLOAD_IMAGE);
     }
   };
