+++ src/app/_components/molecules/ModalContent/postFeed/writeContent/WriteContent.style.tsx
